for some reason, extending textview still doesn't call in any of the accessibility action handlers o_O
but I have learned that if I remove all that garbage, it reverts to the unusable state we had before

so........ I don't know where the bug is hiding, but performAccessibilityAction should get called
for registered actions, and it isn't. Maybe I need an explicit delegate (for some reason)?
I need a break from this...

...maybe try making this potato an actual EditText?
or keep it at TextView, but populate setText() or setAccessibilityText() and see if that gets selection
to at least work (even if performAccessibilityAction ain't getting called, that's not the point)